[ feat ] 회원가입 프로필 페이지 왼편 구현
