+++ src/@pages/signupProfilePage.tsx
@@ -1,3 +1,5 @@
+import SignupProfile from "../@components/signUp/signupProfile";
+
 export default function SignupProfilePage() {
-  return <div>signupProfilePage</div>;
+  return <SignupProfile />;
 }
