+++ src/assets/index.ts
@@ -90,6 +90,8 @@ export { ReactComponent as SignupHeaderStep1Ic } from "./icon/signupHeaderStep1I
 export { ReactComponent as SignupHeaderStep2Ic } from "./icon/signupHeaderStep2Ic.svg";
 export { ReactComponent as SignupHeaderStep3Ic } from "./icon/signupHeaderStep3Ic.svg";
 export { ReactComponent as SignupHeaderStepBlankIc } from "./icon/signupHeaderStepBlankIc.svg";
+export { ReactComponent as SignupProfileCompleteTextIc } from "./icon/signupProfileCompleteTextIc.svg";
+export { ReactComponent as SignupProfileSkipIc } from "./icon/signupProfileSkipIc.svg";
 export { ReactComponent as SignupResendIc } from "./icon/signupResendIc.svg";
 export { ReactComponent as SignupRoleProducerHoverIc } from "./icon/signupRoleProducerHoverIc.svg";
 export { ReactComponent as SignupRoleProducerIc } from "./icon/signupRoleProducerIc.svg";
